+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Challah Rolls 0.2.0
+
+* Dependency updates. Compatability with Rails 3.2.9 and Challah 0.8.3
+
 ## Challah Rolls 0.1.0
 
 * Initial build. Decoupled from challah gem.